+++ src/services/roster/index.js
@@ -4,6 +4,7 @@ const path = require('path');
 const hooks = require('./hooks');
 const globalHooks = require('../../hooks');
 const oauth2 = require('../oauth2/hooks');
+const { excludeAttributesFromSanitization } = require('../../hooks/sanitizationExceptions');
 
 module.exports = function roster() {
 	const app = this;
@@ -68,6 +69,7 @@ module.exports = function roster() {
 				globalHooks.ifNotLocal(hooks.tokenIsActive),
 				globalHooks.ifNotLocal(hooks.userIsMatching),
 				hooks.stripIframe,
+				excludeAttributesFromSanitization('roster/users/:user/metadata', 'username'),
 			],
 		},
 	};
@@ -173,11 +175,11 @@ module.exports = function roster() {
 						data: {
 							students: users.data.map((user) => ({
 								user_id: user.pseudonym,
-								username: encodeURI(oauth2.getSubject(user.pseudonym, app.settings.services.web)),
+								username: oauth2.getSubject(user.pseudonym, app.settings.services.web),
 							})),
 							teachers: teachers.data.map((user) => ({
 								user_id: user.pseudonym,
-								username: encodeURI(oauth2.getSubject(user.pseudonym, app.settings.services.web)),
+								username: oauth2.getSubject(user.pseudonym, app.settings.services.web),
 							})),
 						},
 					}));
@@ -186,7 +188,11 @@ module.exports = function roster() {
 	};
 	const groupsHooks = {
 		before: {
-			get: [globalHooks.ifNotLocal(hooks.tokenIsActive), hooks.injectOriginToolIds],
+			get: [
+				globalHooks.ifNotLocal(hooks.tokenIsActive),
+				hooks.injectOriginToolIds,
+				excludeAttributesFromSanitization('roster/groups', 'username'),
+			],
 		},
 		after: {
 			get: hooks.groupContainsUser,